+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/view/DomainView.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.streamplatform.streamregistry.core.services.unsecured;
+package com.expediagroup.streamplatform.streamregistry.core.view;
 
 import com.expediagroup.streamplatform.streamregistry.model.Domain;
 import com.expediagroup.streamplatform.streamregistry.model.keys.DomainKey;
@@ -25,7 +25,7 @@ import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
-public class UnsecuredDomainService {
+public class DomainServiceView {
   private final DomainRepository domainRepository;
 
   public Optional<Domain> get(DomainKey key) {
