+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/view/InfrastructureView.java
